requirements.txt
python-dotenv
deepgram-sdk
openai
flask